requirements.txt
fastapi
PyJWT
bcrypt
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
alembic 